githubSorter/requirements.txt
PyGithub==1.54.1
psycopg2==1.3.4